pyproject.toml
[project]
name = "shopping-concierge"
version = "0.1.0"
description = "A shopping concierge agent built with Google's Agent Development Kit (ADK)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "litellm",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
